Improvements in the policy  bot.
- Add the labeler plugin which applies labels to issues and prs based on pattern matching in the
title, body, and labels. This is used to perform initial issue triage based on the content of a
filed issue.

- Improved the syncer plugin to dramatically speed up syncing by querying GitHub for only changed
issues and issue comments since the last time we synced. Sadly, this kind of query is not available
for prs and pr reviews.

- Add support for syncing the set of members in each monitored org.

- Add support for running the bot as a syncer rather than a long-lived bot. This
makes the bot perform a sync operation and immediately quit. This is mighty handy
for testing, or bootstrapping syncing for a new org/repo.

- The bot now syncs the file list associated with each PR so that we can run analysis
against the set of files affected by a PR.

- Add low-level support for keeping track of the project's maintainers. Still need the
analysis to leverage this.

- Remove the nagging stats the bot was keeping. It's not really  useful for anything.

- Finish the cache implementation so it caches on reads too, not just on writes.

- Cleanup the dashboard UI so it's ready to get real content.

- Implement a simple ZenHub client library so we can fetch the pipeline info.