PolicyBot improvements
- Implement the unstaler filter. It removes the stale label and comment if an issue
or PR is manipulated in any way. To be useful, I still need to implement the other
side of this which will mark issues/prs as stale with the Istio policy for such.

- Extend the labeler filter's functionality. It can now match prs & issues that have  a
specific set of labels, in addition to the previous negative matching. And it is now able
to remove labels in addition to being able to add them.