abstract resultgatherer use of GCS into blobstorage (#54)
* abstract resultgatherer use of GCS into blobstorage

This change expands blobstorage interface abstraction to support all
of the operations currently used by resultgatherer. The test for
resultgatherer is still broken without credentials, but in the future
we can fake out GCS if we want to.

* create GCS client in server instead of gcs package

This should allow the gcs package to be used for a fake GCS instance,
by specifying a different endpoint when creating the underlying
GCS client. We don't need to expose these details in the gcs package.

* fix resultgatherer_test

* fix import lint error

* undo abstraction change for NewStore