use correct event type for test result filter (#68)
* use correct event type for test result filter

It is now confirmed, using istio/pkg, that Prow uses Github's Status
API, not the Check Runs API. This change adjusts the test result
filter use StatusEvents, which by nature have no PR info associated
with them (statuses are for commits, not PRs). This *should* work for
both commits on PRs and merge commits representing PRs, but the latter
is yet to be tested.