Policy Bot improvements.
- Introduce issue & pr lifecycle tracking functionality. This will add labels
to issues to mark them as needing triage, needing escalation, or being stale.
It'll close issues and prs that have been sitting around for too long.

- Remove stray character that was introduce in a SQL query which was messing up the
members UI.

- Add names to boilerplates in the config such that the names can be used to enhance
the logging done by the pr and issue boilerplate cleaner.

- Add a workaround for bad date display in certain browsers in the UI dashboard.

- Maintain an explicit list of known robot acounts. Robots are not treated like
project members in general, even though they are from GitHub's standpoint.
This is important when we're trying to access whether team members have interacted
with an issue or PR, we don't want to count bots as though they were humans.

- Add support for syncing user data, so we get updated affiliation info. There's
a new cron job that runs weekly for that.

- Simplify startup code for the various command-line tools.

- Clean up cron job-related entrypoints to have a common naming scheme.

- Added cron jobs for label updates, milestone updates, and user data updates.