PolicyBot improvements
- Completely replaced the configuration model. It nows uses an approach akin to the k8s resource model.
Specifically, there is now a series of small named and typed configuration records that together make
up the config of the bot. This makes the growing corpus of YAML easier to manage, and provides a
completely generic and decentralized way to handle config within the bot, which makes it easier
to add new features in the future, and removes a bunch of code throughout the bot's code.

- Added systematic support for dry run mode for all the managers.

- Load user affiliation data from GitHub, like we do for config.
It was previously loaded from the container, requiring a container
update to change the set of users.

- Fixed a variety of bugs discovered along the way.