Add tests for clean regex matches
A precursor to https://github.com/istio/istio/issues/34839, which will
change the format. Now we can make sure it actually works. This fixes a
couple issues with the current regex's, but I don't want to invest much
more time on it since we are going to change them all anyways.