helm: drop secrets (#691)
Currently, we create a policybot secret in the prow cluster, then read
it in the job, pass it as environment variables, read it in the
makefile, pass it back a secret, and deploy that to the cluster.

This is over complex and expands our attack surface - now the same
secret is in two places.

Instead, just store it in the cluster directly. The secret is already
there from previous deployments, so no need to do any manual action to
reconcile this PR.